requirements.txt
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
dj-database-url==0.4.2
Django==2.0.2
django-heroku==0.2.0
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
Mastodon.py==1.2.2
mccabe==0.6.1
pew==1.1.2
pipenv==9.1.0
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
